Move redirect middleware to shared
