--- src/Provider/Provider.Web/Middleware/RequestMiddlewareExtensions.cs
@@ -1,24 +1,0 @@
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
-
-namespace Esfa.Recruit.Provider.Web.Middleware
-{
-    public static class RequestMiddlewareExtensions
-    {
-        public static IApplicationBuilder UseRootRedirect(this IApplicationBuilder builder, string redirectUrl)
-        {
-            return builder.Use(async (context, next) =>
-            {
-                if (context.Request.Path == "/")
-                {
-                    context.Response.Redirect(redirectUrl, true);
-                    await Task.CompletedTask;
-                }
-                else
-                {
-                    await next.Invoke();
-                }
-            });
-        }
-    }
-}
